Move the helper functions for /assemble to assemble-utils
The main assemble route module contained a number of helper functions for the `/assemble` route.

Many or all of these helper functions will be useful for the `/preview` route, so this moves them to the `assemble-utils` module for reuse.

Part of https://github.com/CCALI/a2jdat/issues/94
